Common test helpers moved to base TestCase
