+++ composer.json
@@ -12,6 +12,9 @@
     "autoload": {
         "psr-4": { "React\\Dns\\": "src" }
     },
+    "autoload-dev": {
+        "psr-4": { "React\\Tests\\Dns\\": "tests" }
+    },
     "extra": {
         "branch-alias": {
             "dev-master": "0.4-dev"
